tests/wpt/web-platform-tests/workers/constructors/SharedWorker/port-onmessage.html

<!doctype html>
<title>worker.port.onmessage</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
(async_test()).step(function() {
  var worker = new SharedWorker('port-onmessage.js', '');
  worker.port.onmessage = this.step_func(function(e) {
    assert_true(e.data);
    this.done();
  });
});
</script>
